+++ adapters/Tapjoy/CHANGELOG.md
@@ -1,5 +1,8 @@
 ## Tapjoy iOS Mediation Adapter Changelog
 
+#### Next Version
+- Updated adapter to use new Tapjoy SDK APIs.
+
 #### [Version 13.2.1.0](https://dl.google.com/googleadmobadssdk/mediation/ios/tapjoy/TapjoyAdapter-13.2.1.0.zip)
 - Verified compatibility with Tapjoy SDK 13.2.1.